Se terminan de realizar las bases de lo que pidio el porfe, back end terminado, ahora solo falta corregir estilos en os pdf

+++ database/migrations/2025_08_21_032000_create_product_production_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('production_id')->constrained('production')->onDelete('cascade');
             $table->foreignId('product_id')->constrained('product')->onDelete('cascade');
-            $table->decimal('quantity_produced', 10, 1);
+            $table->integer('quantity_produced');
             $table->decimal('profit_margin_porcentage', 10, 1);
             $table->timestamps();
         });
